Move implementation out of spi package
